+++ docs/src/shared/components/ReactMD/tabs/SimpleTabs.jsx
@@ -13,7 +13,7 @@ export default class SimpleTabs extends PureComponent {
 
   componentWillUnmount() {
     if (this._timeout) {
-      clearTimeout(this._timeut);
+      clearTimeout(this._timeout);
     }
   }