+++ src/js/Autocompletes/__tests__/Autocomplete.js
@@ -8,6 +8,7 @@ import { TAB, DOWN, UP, ENTER, SPACE } from '../../constants/keyCodes';
 import Autocomplete from '../Autocomplete';
 import Menu from '../../Menus/Menu';
 import TextField from '../../TextFields/TextField';
+import List from '../../Lists/List';
 import ListItem from '../../Lists/ListItem';
 
 const DATA = [
@@ -429,5 +430,25 @@ describe('Autocomplete', () => {
     autocomplete.setProps({ data });
     expect(autocomplete.state('matches')).toBe(data);
     expect(autocomplete.state('visible')).toBe(true);
+  });
+
+  
+
+  it('should be renderable inside of a ListItem', () => {
+    const autocomplete = <Autocomplete id="inside-list" data={[]} />;
+
+    let error = false;
+    try {
+      const list = mount(
+        <List>
+          <ListItem primaryText={autocomplete} />
+        </List>
+      );
+      list.find(Autocomplete).simulate('click');
+    } catch (e) {
+      error = true;
+    }
+
+    expect(error).toBe(false);
   });
 });