+++ src/js/Pickers/DatePickerContainer.js
@@ -465,15 +465,17 @@ export default class DatePickerContainer extends PureComponent {
       minDate,
       maxDate,
     } = props;
-    if (typeof props.value !== 'undefined' && props.value !== null) {
-      date = typeof props.value === 'string' ? new Date(props.value) : props.value;
+
+    if (typeof props.value !== 'undefined') {
+      date = this._getDate(props.value);
     } else if (defaultValue) {
-      date = typeof defaultValue === 'string' ? new Date(defaultValue) : defaultValue;
+      date = this._getDate(defaultValue);
       value = typeof defaultValue === 'string'
         ? defaultValue
         : DateTimeFormat(locales, formatOptions).format(defaultValue);
     } else {
       date = new Date();
+      value = '';
     }
 
     date = this._validateDateRange(date, minDate, maxDate);
@@ -483,9 +485,7 @@ export default class DatePickerContainer extends PureComponent {
       : props.defaultCalendarDate;
     let calendarTempDate = date;
     if (typeof defaultCalendarDate !== 'undefined' && !props.value && !props.defaultValue) {
-      calendarTempDate = typeof defaultCalendarDate === 'string'
-        ? new Date(defaultCalendarDate)
-        : defaultCalendarDate;
+      calendarTempDate = this._getDate(defaultCalendarDate);
       date = calendarTempDate;
     } else if (calendarTempDate === null) {
       calendarTempDate = new Date();
@@ -512,7 +512,7 @@ export default class DatePickerContainer extends PureComponent {
     if (this.props.value !== value || !minEqual || !maxEqual) {
       let { calendarDate } = this.state;
       if (typeof value !== 'undefined') {
-        calendarDate = typeof value === 'string' ? new Date(value) : value;
+        calendarDate = this._getDate(value);
       }
 
       calendarDate = this._validateDateRange(calendarDate, minDate, maxDate);
@@ -558,7 +558,17 @@ export default class DatePickerContainer extends PureComponent {
     }
   }
 
-  _setContainer = (container) => {
+  _getDate(value) {
+    if (value === '' || value === null) {
+      return new Date();
+    } else if (typeof value === 'string') {
+      return new Date(value);
+    }
+
+    return value;
+  }
+
+  _setContainer= (container) => {
     this._container = container;
   };