+++ src/js/SelectionControls/SwitchThumb.js
@@ -17,16 +17,8 @@ export default class SwitchThumb extends PureComponent {
     className: PropTypes.string,
     disabled: PropTypes.bool,
     checked: PropTypes.bool,
-    onClick: PropTypes.func.isRequired,
+    onClick: PropTypes.func,
     disabledInteractions: PropTypes.arrayOf(PropTypes.oneOf(['keyboard', 'touch', 'mouse'])),
-  };
-
-  _handleClick = (e) => {
-    // The switch's thumb needs to have an onClick handler to enable keyboard accessibility, however,
-    // this actually triggers two click events since the slider's track needs to be clickable as well.
-    // Stop propagation to prevent the thumb click AND track propagation click.
-    e.stopPropagation();
-    this.props.onClick(e);
   };
 
   render() {
@@ -34,7 +26,6 @@ export default class SwitchThumb extends PureComponent {
     return (
       <AccessibleFakeInkedButton
         {...props}
-        onClick={this._handleClick}
         disabled={disabled}
         disabledInteractions={disabledInteractions || DISABLED_INTERACTIONS}
         inkContainerClassName="md-ink-container--2x"